# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: pgcloud API
  version: "1"
  description: |
    The pgcloud public API. The same API powers the console, CLI, Terraform provider,
    SDKs and AI agents (via MCP in phase 2).

    **Conventions**
    - Base URL: `https://api.pgcloud.example/v1` (domain is an open decision)
    - Auth: `Authorization: Bearer pgc_…` (API token) or a console session JWT
    - Lists are cursor-paginated: `?limit=50&cursor=<id>` → `meta.next_cursor`
    - Mutations accept `Idempotency-Key` (replayed for 24h; same key + different body → 409)
    - Long operations return `202 Accepted` and a resource in a transitional status; poll or use webhooks
    - Money is integer minor units (`amountMinor`) with an explicit `currency` (`USD` | `TRY`)
    - Errors: `{ "error": { "code", "message", "details?" } }`
    - Rate limits: 600 requests per minute per token or session, 120 per minute per IP when anonymous,
      10 sign in attempts per minute per IP. Over the limit: `429 rate_limited` with `Retry-After`
servers:
  - url: http://localhost:4000/v1
    description: local development
security:
  - bearerAuth: []
...
paths:
...
  /deploys:
    get: { tags: [deploys], summary: List Git deployments, parameters: [{ $ref: "#/components/parameters/project" }], responses: { "200": { description: OK } } }
    post:
      tags: [deploys]
      summary: Deploy a Git repository onto a new server
      description: |
        Creates a server whose cloud-init installs Docker, clones the repo, builds (Dockerfile or docker-compose) and serves it on :80.
        Two ways to name the repository: `repoUrl` (public, or private with `gitToken`) returns a per deployment GitHub webhook URL and
        secret (shown once) to add to the repository; `installationId` + `repo` ("owner/name") uses the team's GitHub App installation,
        needs no token, and redeploys on every push through the app webhook (`webhook` is null).
      parameters: [{ $ref: "#/components/parameters/idempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { repoUrl: { type: string, format: uri }, installationId: { type: string }, repo: { type: string, example: acme/app }, branch: { type: string, default: main }, name: { type: string }, port: { type: integer, default: 3000 }, size: { type: string }, project: { type: string }, sshKeys: { type: array, items: { type: string } }, env: { type: object, additionalProperties: { type: string } }, gitToken: { type: string, description: For private repos; stored only in the server's cloud-init } } } } } }
      responses: { "202": { description: Accepted; includes `webhook` (url and secret) for URL based deployments, null for app based ones } }
  /deploys/{id}:
    get: { tags: [deploys], summary: Get a deployment (refreshes status from the server), parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: OK } } }
  /deploys/{id}/redeploy:
    post: { tags: [deploys], summary: Redeploy now, parameters: [{ $ref: "#/components/parameters/id" }], responses: { "202": { description: Accepted } } }
  /deploys/{id}/hook:
    post:
      tags: [deploys]
      security: []
      summary: GitHub push webhook (verified with X-Hub-Signature-256)
      parameters: [{ $ref: "#/components/parameters/id" }, { name: X-Hub-Signature-256, in: header, required: true, schema: { type: string } }, { name: X-GitHub-Event, in: header, schema: { type: string } }]
      responses: { "200": { description: OK }, "401": { $ref: "#/components/responses/Unauthorized" } }
...
  /deploys/{id}/logs:
    get:
      tags: [deploys]
      summary: Tail of the last build log
      description: Fetched live from the server when it is reachable, otherwise the last cached copy. Capped at 32 KB.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { id: { type: string }, status: { type: string }, commit: { type: string, nullable: true }, log: { type: string }, updatedAt: { type: string, format: date-time, nullable: true }, live: { type: boolean, description: true when read from the server just now } } } } } }
...
  /github/connect:
    get:
      tags: [github]
      summary: URL to install the GitHub App for this team
      description: Send the user there. GitHub returns them to the console callback with installation_id and the signed state, which the console posts to /github/installations.
      responses: { "200": { description: OK, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }, "503": { description: Not configured (`github_app_unavailable`) } }
  /github/installations:
    get: { tags: [github], summary: GitHub App installations linked to the team, responses: { "200": { description: OK } } }
    post:
      tags: [github]
      summary: Link an installation after the GitHub redirect
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [installationId, state], properties: { installationId: { type: integer }, state: { type: string } } } } } }
      responses: { "201": { description: Linked }, "401": { description: State invalid or expired } }
  /github/installations/{id}:
    delete: { tags: [github], summary: Unlink and uninstall, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "204": { description: Removed } } }
  /github/installations/{id}/repos:
    get: { tags: [github], summary: Repositories the installation can reach, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "200": { description: OK } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string } }
    project: { name: project, in: query, description: Project id or slug (default `default`), schema: { type: string } }
...
    idempotencyKey: { name: Idempotency-Key, in: header, schema: { type: string, format: uuid } }
  responses:
...
    Unauthorized: { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, unauthorized, forbidden, approval_required, not_found, conflict, invalid_state, quota_exceeded, spend_limit_reached, verification_required, account_suspended, idempotency_key_reused, rate_limited, workflow_unavailable, internal_error]
            message: { type: string }
            details: { type: object, additionalProperties: true }
...